Support logging overhead work on Jira
On our company, overhead work is any work that doesn't have a specific Jira ticket associated to it but rather a generic one (e.g. "Meetings", "1-to-1s", etc).
The company is interested in tracking the time spent on those also. This commit facilitates that by taking advantage of Toggl projects.

Closes #4.